Partition CLI tests across submodules
Moves the 72 #[cfg(test)] tests from src/cli/mod.rs into the
submodule whose subcommand each test exercises:

  setup.rs   15 tests (setup-flag and model-flag cases)
  record.rs  41 tests (record start/stop/toggle/cancel overrides,
                       output-mode flags, profiles, auto-submit,
                       smart auto-submit, shift-enter)
  meeting.rs  4 tests (meeting-start diarization flag matrix)
  root.rs    12 tests (top-level --engine, --driver flags, plus
                       the transcribe subcommand engine override)

mod.rs is now a 22-line barrel that declares submodules and
re-exports the public types; no test code remains there.

Each submodule's test block uses `super::*` plus a glob from
`crate::cli::*` so tests can construct any subcommand variant
regardless of which file owns it (record tests sometimes parse
a full `voxtype record start ...` invocation that returns a
`Cli` whose `command` is `Commands::Record { .. }`).

Behaviour-preserving: cargo test --lib reports the same 770
passing tests. cargo fmt --check, cargo clippy
--all-targets --no-deps -- -D warnings, and the man-page diff
against the dev baseline are all clean.
